+++ pyzbar/pyzbar.py
@@ -9,10 +9,11 @@ from .wrapper import (
     zbar_image_scanner_create, zbar_image_scanner_destroy,
     zbar_image_create, zbar_image_destroy, zbar_image_set_format,
     zbar_image_set_size, zbar_image_set_data, zbar_scan_image,
-    zbar_image_first_symbol, zbar_symbol_get_data_length, zbar_symbol_get_data,
+    zbar_image_first_symbol, zbar_symbol_get_data_length,
+    zbar_symbol_get_data, zbar_symbol_get_orientation,
     zbar_symbol_get_loc_size, zbar_symbol_get_loc_x, zbar_symbol_get_loc_y,
-    zbar_symbol_get_quality, zbar_symbol_next, ZBarConfig, ZBarSymbol,
-    EXTERNAL_DEPENDENCIES
+    zbar_symbol_get_quality, zbar_symbol_next, ZBarConfig, ZBarOrientation,
+    ZBarSymbol, EXTERNAL_DEPENDENCIES,
 )
 
 __all__ = [
@@ -20,7 +21,7 @@ __all__ = [
 ]
 
 
-Decoded = namedtuple('Decoded', 'data type rect polygon quality')
+Decoded = namedtuple('Decoded', 'data type rect polygon quality orientation')
 
 # ZBar's magic 'fourcc' numbers that represent image formats
 _FOURCC = {
@@ -119,13 +120,14 @@ def _decode_symbols(symbols):
             )
             for index in _RANGEFN(zbar_symbol_get_loc_size(symbol))
         )
-
+        orientation = ZBarOrientation(zbar_symbol_get_orientation(symbol)).name
         yield Decoded(
             data=data,
             type=symbol_type,
             rect=bounding_box(polygon),
             polygon=polygon,
-            quality=quality
+            orientation=orientation,
+            quality=quality,
         )