+++ pyzbar/tests/test_pyzbar.py
@@ -40,6 +40,7 @@ class TestDecode(unittest.TestCase):
             type='CODE128',
             rect=Rect(left=37, top=550, width=324, height=76),
             polygon=[(37, 551), (37, 625), (361, 626), (361, 550)],
+            orientation='UP',
             quality=77,
         ),
         Decoded(
@@ -47,6 +48,7 @@ class TestDecode(unittest.TestCase):
             type='CODE128',
             rect=Rect(left=4, top=0, width=390, height=76),
             polygon=[(4, 1), (4, 75), (394, 76), (394, 0)],
+            orientation='UP',
             quality=77,
         ),
     ]
@@ -57,6 +59,7 @@ class TestDecode(unittest.TestCase):
             type='CODE128',
             rect=Rect(left=4, top=0, width=390, height=75),
             polygon=[(4, 1), (4, 75), (394, 74), (394, 0)],
+            orientation="UP",
             quality=76,
         ),
     ]
@@ -67,6 +70,7 @@ class TestDecode(unittest.TestCase):
             type='QRCODE',
             rect=Rect(left=27, top=27, width=145, height=145),
             polygon=[(27, 27), (27, 172), (172, 172), (172, 27)],
+            orientation='UP',
             quality=1,
         ),
     ]
@@ -78,6 +82,7 @@ class TestDecode(unittest.TestCase):
             type='QRCODE',
             rect=Rect(left=173, top=10, width=205, height=205),
             polygon=[(173, 113), (276, 215), (378, 113), (276, 10)],
+            orientation='UP',
             quality=1,
         ),
         Decoded(
@@ -85,6 +90,7 @@ class TestDecode(unittest.TestCase):
             type='QRCODE',
             rect=Rect(left=32, top=208, width=158, height=158),
             polygon=[(32, 352), (177, 366), (190, 222), (46, 208)],
+            orientation='RIGHT',
             quality=1,
         ),
     ]