+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * #108 Python 3.8, 3.9 and 3.10; drop support for Python 3.4
 * #107 Support SQ Codes; tolerate unrecognised codes
 * #86 Read barcodes containing null characters
+* #84 Barcode orientation
 * #46 Include quality
 
 ### v0.1.8